Add comments with image/svg to index.html
Until the images are implemented via svg, img, or background, all
visual components of the google page are linked in an HTML comment.
This should reduce time spent figuring out the minute details of
the visual resources.

Comments will be removed once the visuals are implemented, as the
URL or code will be available in that implementation.